soc/bcm2835: add SDHCI, SPI, GPCLK drivers and Pi Zero 2 W fixes
Peripheral drivers and fixes for bare-metal use on the BCM2710 (Pi Zero
2 W), beyond the low-level ARM/TrustZone bring-up in 7d523d7:

- sdhci.go: Arasan SDHCI/SDIO host controller (polled, line-reset error
  recovery, CMD52/CMD53) -- drives the onboard CYW43438 wireless chip.
- spi.go: SPI0 master (modes 0-3, polled full-duplex).
- gpclk.go: general-purpose clock generator (GPCLK0-2), used for the
  WLAN 32.768 kHz LPO sleep clock.
- mailbox.go: use a plain (non-aliased) message buffer -- the ARM-side
  uncached SDRAM alias only exists on the original BCM2835, so on
  BCM2836/7 the VideoCore read zeros; flush caches around the exchange.
- framebuffer: single-transaction Setup() (size + depth + alloc + pitch).
- miniuart.go: add Rx/Read (the driver was TX-only).
- timer.go: SetAlarm divides in float64 so non-integer timer multipliers
  (19.2 MHz -> 52.083) are not truncated.
- nanotime split (linknanotime tag), bcm2835 float64 arithmetic, and a
  CNTHCTL PL1 timer grant in the HYP-exit path.

Hardware-verified on a Pi Zero 2 W.

+++ arm/timer.go
@@ -108,7 +108,10 @@ func (cpu *CPU) SetAlarm(ns int64) {
 		return
 	}
 
-	set := uint64(ns) / uint64(cpu.TimerMultiplier)
+	// Dividing by a truncated multiplier would scale the absolute deadline
+	// itself: with e.g. a 19.2 MHz counter (multiplier 52.083), truncation
+	// to 52 makes alarms fire late by 0.16% of the time since boot.
+	set := uint64(float64(ns) / cpu.TimerMultiplier)
 	now := read_cntpct()
 	cnt := set - now
